@@ -27,47 +27,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.4.4",
-    "@angular/cdk": "2.0.0-beta.12",
-    "@angular/common": "4.4.4",
-    "@angular/compiler": "4.4.4",
-    "@angular/core": "4.4.4",
-    "@angular/flex-layout": "2.0.0-beta.9",
-    "@angular/forms": "4.4.4",
-    "@angular/http": "4.4.4",
-    "@angular/material": "2.0.0-beta.12",
-    "@angular/platform-browser": "4.4.4",
-    "@angular/platform-browser-dynamic": "4.4.4",
-    "@angular/router": "4.4.4",
+    "@angular/animations": "5.1.1",
+    "@angular/cdk": "5.0.1",
+    "@angular/common": "5.1.1",
+    "@angular/compiler": "5.1.1",
+    "@angular/core": "5.1.1",
+    "@angular/flex-layout": "2.0.0-beta.12",
+    "@angular/forms": "5.1.1",
+    "@angular/http": "5.1.1",
+    "@angular/material": "5.0.1",
+    "@angular/platform-browser": "5.1.1",
+    "@angular/platform-browser-dynamic": "5.1.1",
+    "@angular/router": "5.1.1",
     "@nationallibraryofnorway/ngx-mime": "../dist/",
     "classlist.js": "^1.1.20150312",
     "core-js": "^2.4.1",
     "d3": "^4.10.0",
     "openseadragon": "2.3.1",
-    "rxjs": "^5.1.0",
+    "rxjs": "5.5.2",
     "web-animations-js": "^2.2.5",
-    "zone.js": "^0.8.4"
+    "zone.js": "0.8.14"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.5",
-    "@angular/compiler-cli": "4.4.4",
-    "@angular/language-service": "4.4.4",
-    "@types/chai": "3.5.2",
-    "@types/chai-as-promised": "0.0.30",
-    "@types/cucumber": "2.0.1",
-    "@types/fs-extra": "3.0.2",
-    "@types/jasmine": "2.5.45",
+    "@angular/cli": "1.6.1",
+    "@angular/compiler-cli": "5.1.1",
+    "@angular/language-service": "5.0.2",
+    "@types/chai": "4.0.4",
+    "@types/chai-as-promised": "7.1.0",
+    "@types/cucumber": "2.0.4",
+    "@types/fs-extra": "4.0.3",
+    "@types/jasmine": "~2.5.53",
     "@types/node": "~6.0.60",
     "bluebird": "^3.5.0",
     "body-parser": "^1.18.2",
-    "express": "^4.16.2",
-    "chai": "^3.5.0",
-    "chai-as-promised": "^6.0.0",
-    "codelyzer": "~3.0.1",
+    "chai": "4.1.2",
+    "chai-as-promised": "7.1.1",
+    "chromedriver": "2.33.2",
+    "codelyzer": "4.0.1",
     "concurrently": "^3.4.0",
-    "cucumber": "2.3.1",
-    "fs-extra": "3.0.1",
-    "jasmine-core": "~2.6.2",
+    "cucumber": "3.1.0",
+    "express": "^4.16.2",
+    "fs-extra": "4.0.1",
+    "geckodriver": "1.9.0",
+    "jasmine-core": "2.8.0",
     "jasmine-spec-reporter": "~4.1.0",
     "jsonfile": "3.0.0",
     "karma": "~1.7.0",
@@ -77,14 +79,12 @@
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "mountebank": "^1.11.0",
-    "multiple-cucumber-html-reporter": "1.1.0",
-    "protractor": "~5.1.2",
-    "protractor-cucumber-framework": "3.1.0",
+    "protractor": "5.2.0",
+    "protractor-cucumber-framework": "4.1.1",
+    "protractor-multiple-cucumber-html-reporter-plugin": "^1.2.3",
     "rimraf": "^2.5.4",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3",
-    "chromedriver": "2.33.1",
-    "geckodriver": "1.9.0"
+    "ts-node": "3.3.0",
+    "tslint": "5.7.0",
+    "typescript": "2.5.3"
   }
 }
