@@ -46,39 +46,39 @@
     "release": "sh scripts/release/release.sh"
   },
   "peerDependencies": {
-    "@angular/core": ">=4.4.3 <5.0.0",
-    "@angular/material": ">=2.0.0 <3.0.0 || >=2.0.0-beta.12 <3.0.0",
-    "@angular/cdk": ">=2.0.0 <3.0.0 || >=2.0.0-beta.12 <3.0.0",
-    "@angular/flex-layout": ">=2.0.0 <3.0.0 || >=2.0.0-beta.9 <3.0.0",
-    "openseadragon": ">=2.3.1 <3.0.0",
-    "d3": ">=4.1.0 <5.0.0"
+    "@angular/cdk": ">=5.0.0 <6.0.0",
+    "@angular/core": ">=5.0.0 <6.0.0",
+    "@angular/flex-layout": ">=2.0.0 <3.0.0 || >=2.0.0-beta.12 <3.0.0",
+    "@angular/material": ">=5.0.0 <6.0.0",
+    "d3": ">=4.1.0 <5.0.0",
+    "openseadragon": ">=2.3.1 <3.0.0"
   },
   "devDependencies": {
-    "@angular/animations": "4.4.4",
-    "@angular/cdk": "2.0.0-beta.12",
-    "@angular/common": "4.4.4",
-    "@angular/compiler": "4.4.4",
-    "@angular/compiler-cli": "4.4.4",
-    "@angular/core": "4.4.4",
-    "@angular/flex-layout": "2.0.0-beta.9",
-    "@angular/forms": "4.4.4",
-    "@angular/http": "4.4.4",
-    "@angular/material": "2.0.0-beta.12",
-    "@angular/platform-browser": "4.4.4",
-    "@angular/platform-browser-dynamic": "4.4.4",
-    "@angular/router": "4.4.4",
-    "hammerjs": "2.0.8",
-    "openseadragon": "2.3.1",
-    "@types/jasmine": "2.5.54",
+    "@angular/animations": "5.1.1",
+    "@angular/cdk": "5.0.1",
+    "@angular/common": "5.1.1",
+    "@angular/compiler": "5.1.1",
+    "@angular/compiler-cli": "5.1.1",
+    "@angular/core": "5.1.1",
+    "@angular/flex-layout": "2.0.0-beta.12",
+    "@angular/forms": "5.1.1",
+    "@angular/http": "5.1.1",
+    "@angular/language-service": "5.1.1",
+    "@angular/material": "5.0.1",
+    "@angular/platform-browser": "5.1.1",
+    "@angular/platform-browser-dynamic": "5.1.1",
+    "@angular/router": "5.1.1",
     "@types/d3": "^4.10.0",
+    "@types/jasmine": "2.5.54",
     "@types/node": "^6.0.46",
     "camelcase": "^4.0.0",
-    "codelyzer": "^3.1.1",
+    "codelyzer": "4.0.1",
     "concurrently": "3.4.0",
     "core-js": "^2.4.1",
     "d3": "^4.10.0",
     "execa": "^0.6.3",
     "glob": "^7.1.1",
+    "hammerjs": "2.0.8",
     "jasmine-core": "^2.8.0",
     "karma": "^1.5.0",
     "karma-chrome-launcher": "^2.0.0",
@@ -88,20 +88,23 @@
     "karma-jasmine": "^1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-remap-coverage": "^0.1.4",
-    "lite-server": "^2.2.2",
+    "lite-server": "2.2.2",
     "node-sass": "^4.5.2",
+    "openseadragon": "2.3.1",
     "rimraf": "^2.6.1",
-    "rollup": "^0.42.0",
-    "rollup-plugin-sourcemaps": "^0.4.1",
+    "rollup": "0.42.0",
+    "rollup-plugin-commonjs": "^8.0.2",
+    "rollup-plugin-node-resolve": "^3.0.0",
+    "rollup-plugin-sourcemaps": "0.4.2",
     "rollup-plugin-uglify": "^2.0.1",
-    "rxjs": "5.4.3",
+    "rxjs": "5.5.2",
     "standard-version": "^4.0.0",
-    "systemjs": "^0.19.40",
+    "systemjs": "0.20.19",
     "tslib": "1.7.1",
-    "tslint": "4.5.1",
-    "typescript": "^2.2.1",
+    "tslint": "5.8.0",
+    "typescript": "2.5.3",
     "uglify-es": "^3.0.18",
     "web-animations-js": "^2.3.1",
-    "zone.js": "0.8.16"
+    "zone.js": "0.8.18"
   }
 }
