@@ -1,18 +1,19 @@
 import { Injectable } from '@angular/core';
-import { BehaviorSubject } from 'rxjs/Rx';
+import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { Observable } from 'rxjs/Observable';
+import { distinctUntilChanged } from 'rxjs/operators/distinctUntilChanged';
+
 import { PageRects } from './../models/page-rects';
 import { ViewerLayout } from '../models/viewer-layout';
 import { Point } from './../models/point';
 import { Rect } from './../models/rect';
-
 
 @Injectable()
 export class PageService {
 
   private _numberOfPages: number;
   private _currentNumberOfPages: BehaviorSubject<number> = new BehaviorSubject(0);
-  private _currentPage: BehaviorSubject<number> = new BehaviorSubject(0);
+  protected _currentPage: BehaviorSubject<number> = new BehaviorSubject(0);
 
   private tileRects: Rect[];
   private pageRects: PageRects = new PageRects();
@@ -49,11 +50,13 @@ export class PageService {
   }
 
   get onPageChange(): Observable<number> {
-    return this._currentPage.asObservable().distinctUntilChanged();
+    return this._currentPage.asObservable()
+      .pipe(distinctUntilChanged());
   }
 
   get onNumberOfPagesChange(): Observable<number> {
-    return this._currentNumberOfPages.asObservable().distinctUntilChanged();
+    return this._currentNumberOfPages.asObservable()
+      .pipe(distinctUntilChanged());
   }
 
   get numberOfPages(): number {
