@@ -1,5 +1,5 @@
 import { defineSupportCode } from 'cucumber';
-import { expect } from '../helpers/chai-imports';
+import { expect } from 'chai';
 
 import { ContentSearchPage } from './../pages/content-search.po';
 import { ViewerPage, Point } from '../pages/viewer.po';
@@ -7,26 +7,29 @@ import { ViewerPage, Point } from '../pages/viewer.po';
 defineSupportCode(function ({ Given, When, Then }) {
   const page = new ViewerPage();
   const contentSearchPage = new ContentSearchPage();
+  let selectedHitIndex: number;
 
-  Given(/^the search dialog is open$/, async () => {
+  Given('the search dialog is open', async () => {
     await page.openContentSearchDialog();
   });
 
-  Given(/^the user has selected the second hit$/, async () => {
-    await selectHit(1);
+  Given('the user has search for the word {string}', async (term: string) => {
+    await search(term);
   });
 
-  When(/^the user search for the word "(.*)"$/, async (term: string) => {
-    await page.openContentSearchDialog();
-    await contentSearchPage.setSearchTerm(term);
-    await page.waitForAnimation();
+  Given('the user has selected the {word} hit', async (hit: string) => {
+    await selectHit(hit);
   });
 
-  When(/^the user selects the first hit$/, async () => {
-    await selectHit(0);
+  When('the user search for the word {string}', async (term: string) => {
+    await search(term);
   });
 
-  When(/^the user select the (.*) hit button$/, async (action: string) => {
+  When('the user selects the {word} hit', async (hit: string) => {
+    await selectHit(hit);
+  });
+
+  When('the user select the {word} hit button', async (action: string) => {
     let button;
     if (action === 'previous') {
       button = await contentSearchPage.previousButton();
@@ -39,19 +42,35 @@ defineSupportCode(function ({ Given, When, Then }) {
     await page.waitForAnimation();
   });
 
-  Then(/^there are (.*) results found$/, async (numberOfHits: string) => {
+  When('the user closes the search dialog', async () => {
+    const closeButton = await contentSearchPage.closeButton();
+    await closeButton.click();
+    await page.waitForAnimation();
+  });
+
+  When('the user opens the search dialog', async () => {
+    await page.openContentSearchDialog();
+    await page.waitForAnimation();
+  });
+
+  When('the user click the search inputs clear button', async () => {
+    const clearButton = await contentSearchPage.clearInputButton();
+    await clearButton.click();
+  });
+
+  Then('there are {word} results found', async (numberOfHits: string) => {
     const expected = numberOfHits === 'no' ? 0 : parseInt(numberOfHits, 8);
     const hits = await contentSearchPage.getNumberOfHits();
     expect(hits).to.equal(expected);
   });
 
-  Then(/^the word "(.*)" should be highlighted$/, async (term: string) => {
+  Then('the word {string} should be highlighted', async (term: string) => {
     const hits = await contentSearchPage.getHits();
     const firstHit = await hits[0].getAttribute('innerHTML');
     expect(firstHit).to.contains(`${term} </em>`);
   });
 
-  Then(/^the page with hit number (.*) should be displayed$/, async (hit) => {
+  Then('the page with hit number {word} should be displayed', async (hit) => {
     const currentPageString = await page.getCurrentPageString();
     if (hit === 1) {
       expect(currentPageString.includes('25')).to.eql(true);
@@ -60,30 +79,69 @@ defineSupportCode(function ({ Given, When, Then }) {
     }
   });
 
-  Then(/^all highlighting should be removed$/, async () => {
+  Then('all highlighting should be removed', async () => {
     const hits = await contentSearchPage.getHighlighted();
     expect(hits.length).to.equals(0);
   });
 
-  Then(/^the search result toolbar should be removed$/, async () => {
+  Then('the search result toolbar should be removed', async () => {
     const el = await contentSearchPage.contentSearchNavigatorToolbar();
-    expect(el.isPresent()).to.eql({});
+    expect(await el.isPresent()).to.equal(false);
   });
 
-  Then(/^the Search dialog should be closed$/, async () => {
+  Then('the Search dialog should be {word}', async (state: string) => {
     const isOpen = await contentSearchPage.isOpen();
-    expect(isOpen).to.equal(false);
+    const expectedState = state === 'closed' ? false : true;
+    expect(isOpen).to.equal(expectedState);
   });
 
-  Then(/^hit number (.*) should be marked$/, async (hitIndex: number) => {
-    const isSelected: boolean = await contentSearchPage.isSelected(hitIndex);
+  Then('the hit should be marked', async () => {
+    const isSelected: boolean = await contentSearchPage.hitIsSelected(selectedHitIndex);
     expect(isSelected).to.equal(true);
   });
 
-  async function selectHit(selected: number) {
+  Then('the hit should be visible', async () => {
+    const isVisible: boolean = await contentSearchPage.hitIsVisible(selectedHitIndex);
+    expect(isVisible).to.equal(true);
+  });
+
+  Then('the search query should be empty', async () => {
+    const searchTerm = await contentSearchPage.searchTerm();
+    expect(searchTerm).to.equals('');
+  });
+
+  async function search(term: string) {
+    await page.openContentSearchDialog();
+    await contentSearchPage.setSearchTerm(term);
+    await page.waitForAnimation();
+  }
+
+  async function selectHit(hit: string) {
+    const selected = await hitStringToHitIndex(hit);
     const hits = await contentSearchPage.getHits();
     const first = hits[selected];
     await first.click();
     await page.waitForAnimation();
+    selectedHitIndex = selected;
   }
+
+  async function hitStringToHitIndex(hit: string): Promise<number> {
+    let index: number;
+    if ('first' === hit) {
+      index = 0;
+    } else if ('second' === hit) {
+      index = 1;
+    } else if ('last' === hit) {
+      const hits = await contentSearchPage.getHits();
+      index = hits.length - 1;
+    } else {
+      try {
+        index = parseInt(hit, 10);
+      } catch (e) {
+        throw new Error(`Unrecognized value "${hit}`);
+      }
+    }
+    return index;
+  }
+
 });
