@@ -1,7 +1,8 @@
-import { ViewerPage } from '../pages/viewer.po';
 import { defineSupportCode } from 'cucumber';
-import { expect } from '../helpers/chai-imports';
+import { expect } from 'chai';
 import { browser, by } from 'protractor';
+
+import { ViewerPage } from '../pages/viewer.po';
 import { Utils } from '../helpers/utils';
 
 defineSupportCode(function ({ Given, When, Then }) {
@@ -10,8 +11,7 @@ defineSupportCode(function ({ Given, When, Then }) {
   // 300ms is the time the animation is set up to use but we need some extra time
   const switchAnimationTime = 1700;
 
-
-  Given(/^the layout is two-page$/, async () => {
+  Given('the layout is two-page', async () => {
     const btn = await page.getTwoPageButton();
     // Button is present, so click to switch to two-page
     if (btn) {
@@ -21,7 +21,7 @@ defineSupportCode(function ({ Given, When, Then }) {
     expect(await page.isTwoPageView()).to.equal(true);
   });
 
-  Given(/^the layout is one-page$/, async () => {
+  Given('the layout is one-page', async () => {
     const btn = await page.getOnePageButton();
     // Button is present, so switch to one-page
     if (btn) {
@@ -32,7 +32,7 @@ defineSupportCode(function ({ Given, When, Then }) {
     expect(await page.isOnePageView()).to.equal(true);
   });
 
-  Then(/^only the cover page is displayed/, async () => {
+  Then('only the cover page is displayed', async () => {
     await page.waitForAnimation(switchAnimationTime);
     const visiblePages = await page.visiblePages();
     // Firste page visible
@@ -42,7 +42,7 @@ defineSupportCode(function ({ Given, When, Then }) {
     expect(visiblePages).to.not.include(true);
   });
 
-  Then(/^page 2 and 3 are displayed/, async () => {
+  Then('page 2 and 3 are displayed', async () => {
     await page.waitForAnimation(switchAnimationTime);
 
     const visiblePages = await page.visiblePages();
@@ -55,7 +55,7 @@ defineSupportCode(function ({ Given, When, Then }) {
     expect(visiblePages).to.not.include(true);
   });
 
-  Then(/^only page 2 is displayed/, async () => {
+  Then('only page 2 is displayed', async () => {
     await page.waitForAnimation(switchAnimationTime);
     const visiblePages = await page.visiblePages();
     // Second page visible
