@@ -1,14 +1,14 @@
 import { defineSupportCode } from 'cucumber';
-import { expect } from '../helpers/chai-imports';
+import { expect } from 'chai';
 
 import { ViewerPage } from './../pages/viewer.po';
 import { MetadataPage } from './../pages/metadata.po';
 
-defineSupportCode(function ({ Given, Then }) {
+defineSupportCode(function ({ Given, When, Then }) {
   const viewer = new ViewerPage();
   const metadata = new MetadataPage();
 
-  Then(/the attribution must be shown$/, async () => {
+  Then('the attribution must be shown', async () => {
     const pageAttribution = await viewer.getAttribution();
     await viewer.openContentsDialog();
     const metadataAttribution = await metadata.getAttribution();
@@ -17,7 +17,7 @@ defineSupportCode(function ({ Given, Then }) {
     expect(await metadataAttribution.getText()).to.eql('This is a test attribution');
   });
 
-  Then(/the license must be shown as hyperlinks$/, async () => {
+  Then('the license must be shown as hyperlinks', async () => {
     await viewer.openContentsDialog();
     const license = await metadata.getLicense();
 
