@@ -16,14 +16,19 @@
       // angular bundles
       '@angular/core': 'npm:@angular/core/bundles/core.umd.js',
       '@angular/common': 'npm:@angular/common/bundles/common.umd.js',
-      '@angular/common/http': 'npm:@angular/common/bundles/common-http.umd.js',
       '@angular/compiler': 'npm:@angular/compiler/bundles/compiler.umd.js',
       '@angular/platform-browser': 'npm:@angular/platform-browser/bundles/platform-browser.umd.js',
       '@angular/platform-browser-dynamic': 'npm:@angular/platform-browser-dynamic/bundles/platform-browser-dynamic.umd.js',
+      '@angular/flex-layout': 'npm:@angular/flex-layout/bundles/flex-layout.umd.js',
       '@angular/http': 'npm:@angular/http/bundles/http.umd.js',
+      '@angular/common/http': 'npm:@angular/common/bundles/common-http.umd.js',
       '@angular/router': 'npm:@angular/router/bundles/router.umd.js',
       '@angular/forms': 'npm:@angular/forms/bundles/forms.umd.js',
       '@angular/material': 'npm:@angular/material/bundles/material.umd.js',
+      '@angular/animations': 'npm:@angular/animations/bundles/animations.umd.js',
+      '@angular/animations/browser': 'npm:@angular/animations/bundles/animations-browser.umd.js',
+      '@angular/platform-browser/animations': 'npm:@angular/platform-browser/bundles/platform-browser-animations.umd.js',
+
       '@angular/cdk': 'npm:@angular/cdk/bundles/cdk.umd.js',
       '@angular/cdk/a11y': 'npm:@angular/cdk/bundles/cdk-a11y.umd.js',
       '@angular/cdk/bidi': 'npm:@angular/cdk/bundles/cdk-bidi.umd.js',
@@ -38,18 +43,15 @@
       '@angular/cdk/scrolling': 'npm:@angular/cdk/bundles/cdk-scrolling.umd.js',
       '@angular/cdk/stepper': 'npm:@angular/cdk/bundles/cdk-stepper.umd.js',
       '@angular/cdk/table': 'npm:@angular/cdk/bundles/cdk-table.umd.js',
-      '@angular/cdk/testing': 'npm:@angular/cdk/bundles/cdk-testing.umd.js',
-      '@angular/animations': 'npm:@angular/animations/bundles/animations.umd.js',
-      '@angular/animations/browser': 'npm:@angular/animations/bundles/animations-browser.umd.js',
-      '@angular/platform-browser/animations': 'npm:@angular/platform-browser/bundles/platform-browser-animations.umd.js',
-      '@angular/flex-layout' : 'npm:@angular/flex-layout/bundles/flex-layout.umd.js',
+      '@angular/cdk/accordion': 'npm:@angular/cdk/bundles/cdk-accordion.umd.js',
+      '@angular/cdk/layout': 'npm:@angular/cdk/bundles/cdk-layout.umd.js',
 
       // other libraries
-      'tslib': 'npm:tslib/tslib.js',
       'rxjs': 'npm:rxjs',
       'hammerjs': 'npm:hammerjs/hammer.js',
       'openseadragon': 'npm:openseadragon/build/openseadragon/openseadragon.min.js',
-      'd3': 'npm:d3/build/d3.min.js'
+      'd3': 'npm:d3/build/d3.min.js',
+      'tslib': 'npm:tslib/tslib.js'
     },
     // packages tells the System loader how to load when no filename and/or no extension
     packages: {
@@ -63,6 +65,10 @@
       },
       rxjs: {
         defaultExtension: 'js'
+      },
+      'rxjs/operators': {
+        defaultExtension: 'js',
+        main: 'index'
       },
       'ngx-mime': {
         main: 'index.js',
