@@ -1,5 +1,6 @@
 import { Component, OnInit, Input, ChangeDetectionStrategy, ChangeDetectorRef, OnDestroy } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
+import { Subject } from 'rxjs/Subject';
+import { takeUntil } from 'rxjs/operators/takeUntil';
 
 import { MimeViewerIntl } from './../../core/intl/viewer-intl';
 import { IiifManifestService } from './../../core/iiif-manifest-service/iiif-manifest-service';
@@ -13,7 +14,7 @@ import { Manifest } from './../../core/models/manifest';
 })
 export class MetadataComponent implements OnInit, OnDestroy {
   public manifest: Manifest;
-  private subscriptions: Array<Subscription> = [];
+  private destroyed: Subject<void> = new Subject();
 
   constructor(
     public intl: MimeViewerIntl,
@@ -21,17 +22,20 @@ export class MetadataComponent implements OnInit, OnDestroy {
     private iiifManifestService: IiifManifestService) { }
 
   ngOnInit() {
-    this.subscriptions.push(this.iiifManifestService.currentManifest
+    this.iiifManifestService
+      .currentManifest
+      .pipe(
+        takeUntil(this.destroyed)
+      )
       .subscribe((manifest: Manifest) => {
         this.manifest = manifest;
         this.changeDetectorRef.markForCheck();
-      }));
+      });
   }
 
   ngOnDestroy() {
-    this.subscriptions.forEach((subscription: Subscription) => {
-      subscription.unsubscribe();
-    });
+    this.destroyed.next();
+    this.destroyed.complete();
   }
 
 }
