@@ -3,8 +3,8 @@ import { Point } from '../models/point';
 export class ZoomUtils {
   /**
    *
-   * @param {Point} in OSD-viewport-coordinates
-   * @param {Rect} pageBounds
+   * @param Point in OSD-viewport-coordinates
+   * @param Rect pageBounds
    */
   static constrainPositionToPage(point: Point, pageBounds: Rect): Point {
     if (point.x < pageBounds.x) {
