@@ -6,6 +6,7 @@ import { SharedModule } from './shared/shared.module';
 import { ContentsDialogModule } from './contents-dialog/contents-dialog.module';
 import { AttributionDialogModule } from './attribution-dialog/attribution-dialog.module';
 import { ContentSearchDialogModule } from './content-search-dialog/content-search-dialog.module';
+import { PageDialogModule } from './page-dialog/page-dialog.module';
 import { ViewerComponent } from './viewer/viewer.component';
 import { ViewerHeaderComponent } from './viewer/viewer-header/viewer-header.component';
 import { ViewerFooterComponent } from './viewer/viewer-footer/viewer-footer.component';
@@ -14,7 +15,6 @@ import { PageNavigatorComponent } from './viewer/viewer-footer/page-navigator/pa
 import { ViewerSpinnerComponent } from './viewer/viewer-spinner/viewer-spinner.component';
 import { OsdToolbarComponent } from './viewer/osd-toolbar/osd-toolbar.component';
 
-import './rxjs-extension';
 import 'openseadragon';
 import 'd3';
 
@@ -33,7 +33,8 @@ import 'd3';
     SharedModule,
     ContentsDialogModule,
     AttributionDialogModule,
-    ContentSearchDialogModule
+    ContentSearchDialogModule,
+    PageDialogModule
   ],
   exports: [
     ViewerComponent
