@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule }   from '@angular/forms';
+import { FormsModule, ReactiveFormsModule }   from '@angular/forms';
 import { FlexLayoutModule } from '@angular/flex-layout';
 
 import { MimeMaterialModule } from './mime-material.module';
@@ -9,15 +9,17 @@ import { SpinnerService } from '../core/spinner-service/spinner.service';
 @NgModule({
   imports: [
     CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
     FlexLayoutModule,
-    MimeMaterialModule,
-    FormsModule
+    MimeMaterialModule
   ],
   exports: [
     CommonModule,
     FlexLayoutModule,
-    MimeMaterialModule,
-    FormsModule
+    FormsModule,
+    ReactiveFormsModule,
+    MimeMaterialModule
   ],
   providers: [
     SpinnerService
