@@ -18,11 +18,15 @@ import {
 } from '../../../core/navigation/content-search-navigation-service/content-search-navigation.service';
 import { IiifContentSearchServiceStub } from '../../../test/iiif-content-search-service-stub';
 import { ViewerServiceMock } from './../../../test/viewer-service-mock';
+import { Rect } from '../../../core/models/rect';
+import { ViewerLayout } from '../../../core/models/viewer-layout';
 
 describe('ContentSearchNavigatorComponent', () => {
   let component: ContentSearchNavigatorComponent;
   let fixture: ComponentFixture<ContentSearchNavigatorComponent>;
-  let iiifContentSearchServiceStub: IiifContentSearchServiceStub;
+  let iiifContentSearchService: IiifContentSearchServiceStub;
+  let pageService: PageServiceMock;
+
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -35,19 +39,22 @@ describe('ContentSearchNavigatorComponent', () => {
         MimeViewerIntl,
         ContentSearchNavigationService,
         { provide: ViewerService, useClass: ViewerServiceMock },
-        { provide: PageService, useClass: PageServiceMock },
         { provide: IiifContentSearchService, useClass: IiifContentSearchServiceStub },
+        { provide: PageService, useClass: PageServiceMock }
       ]
 
     }).compileComponents();
   }));
 
   beforeEach(() => {
-    iiifContentSearchServiceStub = TestBed.get(IiifContentSearchService);
     fixture = TestBed.createComponent(ContentSearchNavigatorComponent);
     component = fixture.componentInstance;
     component.searchResult = createDefaultData();
-    iiifContentSearchServiceStub._currentSearchResult.next(component.searchResult);
+    iiifContentSearchService._currentSearchResult.next(component.searchResult);
+
+    iiifContentSearchService = TestBed.get(IiifContentSearchService);
+    pageService = TestBed.get(PageService);
+    pageService.addPages(createDefaultTileRects(102), ViewerLayout.TWO_PAGE, true);
     fixture.detectChanges();
   });
 
@@ -74,6 +81,7 @@ describe('ContentSearchNavigatorComponent', () => {
 
       viewerService.setPageChange(4);
       fixture.detectChanges();
+
       fixture.whenStable().then(() => {
         const res = component.goToPreviousHitPage();
         expect(csns.goToPreviousHitPage).toHaveBeenCalledTimes(1);
@@ -88,6 +96,7 @@ describe('ContentSearchNavigatorComponent', () => {
 
       pageService.setPageChange(3);
       fixture.detectChanges();
+
       fixture.whenStable().then(() => {
         const res = component.goToPreviousHitPage();
         expect(csns.goToPreviousHitPage).toHaveBeenCalledTimes(1);
@@ -98,50 +107,203 @@ describe('ContentSearchNavigatorComponent', () => {
   it('should go to first hit if current canvas is 0 and user presses next hit button',
     inject([ViewerService, PageService, ContentSearchNavigationService],
       (viewerService: ViewerServiceMock, pageService: PageServiceMock, csns: ContentSearchNavigationService) => {
-      spyOn(csns, 'goToNextHitPage').and.callThrough();
+        spyOn(csns, 'goToNextHitPage').and.callThrough();
+  });
 
+    it('should go to first hit if user is before first hit and presses next hit button', () => {
+      spyOn(iiifContentSearchService, 'selected');
       pageService.setPageChange(0);
       fixture.detectChanges();
+
       fixture.whenStable().then(() => {
         const res = component.goToNextHitPage();
         expect(csns.goToNextHitPage).toHaveBeenCalledTimes(1);
         expect(csns.getCurrentIndex()).toEqual(0);
       });
-    }));
+    });
 
-  it('should disable previous button if on first hit',
-    inject([ViewerService, PageService], (viewerService: ViewerServiceMock, pageService: PageServiceMock) => {
+    it('should disable previous button if on first hit', () => {
       pageService.setPageChange(2);
       fixture.detectChanges();
       fixture.whenStable().then(() => {
         const button = fixture.debugElement.query(By.css('#footerNavigatePreviousHitButton'));
         expect(button.nativeElement.disabled).toBeTruthy();
       });
-    }));
+    });
 
-  it('should disable next button if on last hit',
-    inject([ViewerService, PageService], (viewerService: ViewerServiceMock, pageService: PageServiceMock) => {
-      pageService.setPageChange(8);
+    it('should disable next button if on last hit', () => {
+      const last = createDefaultData().last();
+      pageService.setPageChange(51);
       fixture.detectChanges();
       fixture.whenStable().then(() => {
         const button = fixture.debugElement.query(By.css('#footerNavigateNextHitButton'));
         expect(button.nativeElement.disabled).toBeTruthy();
       });
-    }));
+
+    });
+
+    it('should go to first hit on left page if user presses previous hit button', () => {
+      spyOn(iiifContentSearchService, 'selected');
+      component.searchResult = createLeftPageHit();
+
+      pageService.setPageChange(2);
+      fixture.detectChanges();
+
+      fixture.whenStable().then(() => {
+        component.goToPreviousHitPage();
+        expect(iiifContentSearchService.selected)
+          .toHaveBeenCalledWith(new Hit({ id: 1, index: 1 }));
+      });
+
+    });
+
+    it('should go to first hit on right page if user presses previous hit button', () => {
+      spyOn(iiifContentSearchService, 'selected');
+      component.searchResult = createRightPageHit();
+      pageService.setPageChange(3);
+      fixture.detectChanges();
+
+      fixture.whenStable().then(() => {
+        component.goToPreviousHitPage();
+        expect(iiifContentSearchService.selected).toHaveBeenCalledWith(new Hit({ id: 1, index: 2 }));
+      });
+
+    });
+
+    it('should skip going to right page when there is hits on both pages when user presses next hit button', () => {
+      spyOn(iiifContentSearchService, 'selected');
+      component.searchResult = createRightPageHit();
+      pageService.setPageChange(0);
+      fixture.detectChanges();
+
+      fixture.whenStable().then(() => {
+        component.goToNextHitPage();
+        expect(iiifContentSearchService.selected).toHaveBeenCalledWith(new Hit({ id: 1, index: 2 }));
+        component.goToNextHitPage();
+        expect(iiifContentSearchService.selected).toHaveBeenCalledWith(new Hit({ id: 3, index: 100 }));
+      });
+
+    });
+  });
+
+  describe('Single page display', () => {
+
+    it('should go to first hit if user presses previous hit button', () => {
+      spyOn(iiifContentSearchService, 'selected');
+      component.searchResult = createSinglePageHit();
+      pageService.reset()
+      pageService.addPages(createDefaultTileRects(102), ViewerLayout.ONE_PAGE, true);
+      pageService.setPageChange(3);
+      fixture.detectChanges();
+
+      fixture.whenStable().then(() => {
+        component.goToPreviousHitPage();
+        expect(iiifContentSearchService.selected).toHaveBeenCalledWith(new Hit({ id: 1, index: 2 }));
+      });
+
+    });
+
+    it('should go to next hit if user presses next hit button', () => {
+      spyOn(iiifContentSearchService, 'selected');
+      pageService.setPageChange(0);
+      fixture.detectChanges();
+
+      fixture.whenStable().then(() => {
+        component.goToNextHitPage();
+        expect(iiifContentSearchService.selected).toHaveBeenCalledWith(new Hit({ id: 1, index: 1 }));
+        component.goToNextHitPage();
+        expect(iiifContentSearchService.selected).toHaveBeenCalledWith(new Hit({ id: 2, index: 8 }));
+      });
+
+    });
+
+  });
+
+  function createDefaultTileRects(size: number): Rect[] {
+    const tileRects: Rect[] = [];
+    for (let i = 0; i < size; i++) {
+      tileRects.push(new Rect());
+    }
+    return tileRects;
+  }
 
   function createDefaultData() {
     const searchResult = new SearchResult();
     searchResult.add(new Hit({
+      id: 1,
+      index: 1
+    }));
+    searchResult.add(new Hit({
+      id: 2,
+      index: 8
+    }));
+    searchResult.add(new Hit({
+      id: 3,
+      index: 10
+    }));
+    searchResult.add(new Hit({
+      id: 4,
+      index: 20
+    }));
+    searchResult.add(new Hit({
+      id: 5,
+      index: 30
+    }));
+    searchResult.add(new Hit({
+      id: 6,
+      index: 40
+    }));
+    searchResult.add(new Hit({
+      id: 7,
+      index: 100
+    }));
+    return searchResult;
+  }
+
+  function createLeftPageHit() {
+    const searchResult = new SearchResult();
+    searchResult.add(new Hit({
+      id: 1,
+      index: 1
+    }));
+    searchResult.add(new Hit({
+      id: 2,
+      index: 1
+    }));
+    searchResult.add(new Hit({
+      index: 2
+    });
+
+    return searchResult;
+  }
+
+  function createRightPageHit() {
+    const searchResult = new SearchResult();
+    searchResult.add(new Hit({
+      id: 1,
       index: 2
     }));
     searchResult.add(new Hit({
-      index: 4
+      id: 2,
+      index: 2
     }));
     searchResult.add(new Hit({
-      index: 4
+      id: 3,
+      index: 100
+    }));
+
+    return searchResult;
+  }
+
+  function createSinglePageHit() {
+    const searchResult = new SearchResult();
+    searchResult.add(new Hit({
+      id: 1,
+      index: 2
     }));
     searchResult.add(new Hit({
-      index: 8
+      id: 2,
+      index: 2
     }));
 
     return searchResult;
@@ -149,27 +311,10 @@ describe('ContentSearchNavigatorComponent', () => {
 
 });
 
-class IiifContentSearchServiceMock {
-  _onChange = new Subject<number>();
-  get onChange(): Observable<number> {
-    return this._onChange.asObservable();
-  }
-
-}
-
-class PageServiceMock {
-
-  pageChanged = new Subject<number>();
-  get onPageChange(): Observable<number> {
-    return this.pageChanged.asObservable();
-  }
+class PageServiceMock extends PageService {
 
   setPageChange(index: number) {
-    this.pageChanged.next(index);
+    this._currentPage.next(index);
   }
 
-  getTileArrayFromPageIndex(index: number): number[] {
-    return [index];
-  }
 }
-
